+++ src/vma/proto/dst_entry_udp.h
@@ -61,7 +61,7 @@ private:
 	const uint32_t m_n_tx_bufs_batch_udp;
 	const bool m_b_tx_nonblocked_eagains;
 	const thread_mode_t	m_thread_mode;
-	const size_t m_n_tx_prefetch_bytes;
+	const uint32_t m_n_tx_prefetch_bytes;
 };
 
 #endif /* DST_ENTRY_UDP_H */
